# app/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "DreamTeam Crowdfunding Platform API",
    "description": "This is the API for the DreamTeam Crowdfunding Platform. It provides endpoints for managing users, contributors, entrepreneurs, startups, and contributions.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000/"
    }
  ],
...
  "paths": {
...
    "/contributors": {
      "post": {
        "tags": [
          "contributors"
        ],
        "summary": "Create a new contributor",
        "description": "Create a new contributor in the system",
        "operationId": "createContributor",
        "requestBody": {
          "description": "Created contributor object",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ContributorInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Contributor"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input"
          }
        }
      }
    },
    "/contributors/{contributorId}": {
      "get": {
        "tags": [
          "contributors"
        ],
        "summary": "Get contributor by ID",
        "description": "Returns a single contributor",
        "operationId": "getContributorById",
        "parameters": [
          {
            "name": "contributorId",
            "in": "path",
            "description": "ID of contributor to return",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Contributor"
                }
              }
            }
          },
          "404": {
            "description": "Contributor not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "put": {
        "tags": [
          "contributors"
        ],
        "summary": "Update an existing contributor",
        "description": "Update an existing contributor by Id",
        "operationId": "updateContributor",
        "parameters": [
          {
            "name": "contributorId",
            "in": "path",
            "description": "ID of contributor to update",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "description": "Updated contributor object",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Contributor"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Contributor"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input"
          },
          "404": {
            "description": "Contributor not found"
          }
        }
      },
      "delete": {
        "tags": [
          "contributors"
        ],
        "summary": "Delete a contributor",
        "description": "Delete a contributor",
        "operationId": "deleteContributor",
        "parameters": [
          {
            "name": "contributorId",
            "in": "path",
            "description": "Contributor id to delete",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation"
          },
          "404": {
            "description": "Contributor not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
    "/regions": {
      "post": {
        "tags": [
          "regions"
        ],
        "summary": "Create a new region",
        "description": "Create a new region in the system",
        "operationId": "createRegion",
        "requestBody": {
          "description": "Created region object",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Region"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Region"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input"
          }
        }
      },
      "get": {
        "tags": [
          "regions"
        ],
        "summary": "List all regions",
        "description": "Returns a list of all regions",
        "operationId": "listRegions",
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Region"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/regions/{regionId}": {
      "get": {
        "tags": [
          "regions"
        ],
        "summary": "Get region by ID",
        "description": "Returns a single region",
        "operationId": "getRegionById",
        "parameters": [
          {
            "name": "regionId",
            "in": "path",
            "description": "ID of region to return",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Region"
                }
              }
            }
          },
          "404": {
            "description": "Region not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "put": {
        "tags": [
          "regions"
        ],
        "summary": "Update an existing region",
        "description": "Update an existing region by Id",
        "operationId": "updateRegion",
        "parameters": [
          {
            "name": "regionId",
            "in": "path",
            "description": "ID of region to update",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "description": "Updated region object",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Region"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Region"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input"
          },
          "404": {
            "description": "Region not found"
          }
        }
      },
      "delete": {
        "tags": [
          "regions"
        ],
        "summary": "Delete a region",
        "description": "Delete a region",
        "operationId": "deleteRegion",
        "parameters": [
          {
            "name": "regionId",
            "in": "path",
            "description": "Region id to delete",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation"
          },
          "404": {
            "description": "Region not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/phones": {
      "post": {
        "tags": [
          "phones"
        ],
        "summary": "Create a new phone",
        "description": "Create a new phone number in the system",
        "operationId": "createPhone",
        "requestBody": {
          "description": "Created phone object",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Phone"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Phone"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input"
          }
        }
      },
      "get": {
        "tags": [
          "phones"
        ],
        "summary": "List all phones",
        "description": "Returns a list of all phone numbers",
        "operationId": "listPhones",
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Phone"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/phones/{phoneId}": {
      "get": {
        "tags": [
          "phones"
        ],
        "summary": "Get phone by ID",
        "description": "Returns a single phone number",
        "operationId": "getPhoneById",
        "parameters": [
          {
            "name": "phoneId",
            "in": "path",
            "description": "ID of phone to return",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Phone"
                }
              }
            }
          },
          "404": {
            "description": "Phone not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "put": {
        "tags": [
          "phones"
        ],
        "summary": "Update an existing phone",
        "description": "Update an existing phone number by Id",
        "operationId": "updatePhone",
        "parameters": [
          {
            "name": "phoneId",
            "in": "path",
            "description": "ID of phone to update",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "description": "Updated phone object",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Phone"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Phone"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input"
          },
          "404": {
            "description": "Phone not found"
          }
        }
      },
      "delete": {
        "tags": [
          "phones"
        ],
        "summary": "Delete a phone",
        "description": "Delete a phone number",
        "operationId": "deletePhone",
        "parameters": [
          {
            "name": "phoneId",
            "in": "path",
            "description": "Phone id to delete",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation"
          },
          "404": {
            "description": "Phone not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "Contributor": {
        "type": "object",
        "properties": {
          "contributor_id": {
            "type": "integer",
            "format": "int64",
            "description": "The auto-generated id of the contributor"
          },
          "gender": {
            "type": "string",
            "enum": [
              "male",
              "female",
              "non_binary"
            ]
          },
          "dob": {
            "type": "string",
            "format": "date-time"
          },
          "passport_id": {
            "type": "integer",
            "format": "int64"
          },
          "agreement_id": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "ContributorInput": {
        "type": "object",
        "required": [
          "gender",
          "dob",
          "passport"
        ],
        "properties": {
          "gender": {
            "type": "string",
            "enum": [
              "male",
              "female",
              "non_binary"
            ]
          },
          "dob": {
            "type": "string",
            "format": "date-time"
          },
          "passport": {
            "$ref": "#/components/schemas/Passport"
          },
          "bankAgreement": {
            "$ref": "#/components/schemas/BankAgreement"
          }
        }
      },
...
      "Passport": {
        "type": "object",
        "required": [
          "series",
          "number"
        ],
        "properties": {
          "series": {
            "type": "string"
          },
          "number": {
            "type": "integer"
          },
          "issue_date": {
            "type": "string",
            "format": "date-time"
          },
          "expiration_date": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "BankAgreement": {
        "type": "object",
        "properties": {
          "document": {
            "type": "object"
          },
          "is_signed": {
            "type": "boolean"
          }
        }
      },
...
      "Region": {
        "type": "object",
        "required": [
          "name"
        ],
        "properties": {
          "region_id": {
            "type": "integer",
            "format": "int64",
            "description": "The auto-generated id of the region"
          },
          "name": {
            "type": "string",
            "description": "The name of the region"
          }
        }
      },
      "Phone": {
        "type": "object",
        "required": [
          "mobile_operator_code",
          "phone_number"
        ],
        "properties": {
          "phone_id": {
            "type": "integer",
            "format": "int64",
            "description": "The auto-generated id of the phone"
          },
          "country_code": {
            "type": "integer",
            "description": "The country code",
            "default": 998
          },
          "mobile_operator_code": {
            "type": "integer",
            "description": "The mobile operator code"
          },
          "phone_number": {
            "type": "integer",
            "description": "The phone number"
          }
        }
      }
    }
  }
}